+++ app/src/main/java/com/decoutkhanqindev/dexreader/domain/repository/network/NetworkRepository.kt
@@ -1,0 +1,7 @@
+package com.decoutkhanqindev.dexreader.domain.repository.network
+
+import kotlinx.coroutines.flow.Flow
+
+interface NetworkRepository {
+  fun observeIsAvailable(): Flow<Boolean>
+}