+++ app/src/main/java/com/decoutkhanqindev/dexreader/domain/usecase/prefs/SaveIsOnboardingCompletedUseCase.kt
@@ -1,11 +1,11 @@
-package com.decoutkhanqindev.dexreader.domain.usecase.settings
+  package com.decoutkhanqindev.dexreader.domain.usecase.prefs
 
-import com.decoutkhanqindev.dexreader.domain.repository.settings.SettingsRepository
+import com.decoutkhanqindev.dexreader.domain.repository.prefs.PrefsRepository
 import com.decoutkhanqindev.dexreader.util.CoroutineHandler.runSuspendResultCatching
 import javax.inject.Inject
 
 class SaveIsOnboardingCompletedUseCase @Inject constructor(
-  private val repository: SettingsRepository,
+  private val repository: PrefsRepository,
 ) {
   suspend operator fun invoke(value: Boolean): Result<Unit> =
     runSuspendResultCatching { repository.saveIsOnboardingCompleted(value) }