+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/common/viewmodels/prefs/PrefsData.kt
@@ -1,10 +1,10 @@
-package com.decoutkhanqindev.dexreader.presentation.screens.common.viewmodels.settings
+package com.decoutkhanqindev.dexreader.presentation.screens.common.viewmodels.prefs
 
 import androidx.compose.runtime.Immutable
 import com.decoutkhanqindev.dexreader.presentation.model.value.settings.ThemeModeValue
 
 @Immutable
-data class SettingsUiState(
+data class PrefsData(
   val isLoading: Boolean = false,
   val appliedThemeOption: ThemeModeValue = ThemeModeValue.DARK,
   val selectedThemeOption: ThemeModeValue = ThemeModeValue.DARK,