+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/settings/SettingsScreen.kt
@@ -11,7 +11,7 @@ import com.decoutkhanqindev.dexreader.R
 import com.decoutkhanqindev.dexreader.presentation.model.value.settings.ThemeModeValue
 import com.decoutkhanqindev.dexreader.presentation.navigation.NavRoute
 import com.decoutkhanqindev.dexreader.presentation.screens.common.base.BaseDetailsScreen
-import com.decoutkhanqindev.dexreader.presentation.screens.common.viewmodels.settings.SettingsViewModel
+import com.decoutkhanqindev.dexreader.presentation.screens.common.viewmodels.prefs.PrefsViewModel
 import com.decoutkhanqindev.dexreader.presentation.screens.settings.components.SettingsContent
 import com.decoutkhanqindev.dexreader.util.NavTransitions.navigateBack
 import com.decoutkhanqindev.dexreader.util.NavTransitions.navigateTo
@@ -19,10 +19,10 @@ import com.decoutkhanqindev.dexreader.util.NavTransitions.navigateTo
 @Composable
 fun SettingsScreen(
   navController: NavHostController,
-  settingsViewModel: SettingsViewModel,
+  prefsViewModel: PrefsViewModel,
   modifier: Modifier = Modifier,
 ) {
-  val uiState by settingsViewModel.uiState.collectAsStateWithLifecycle()
+  val uiState by prefsViewModel.data.collectAsStateWithLifecycle()
 
   BaseDetailsScreen(
     title = stringResource(R.string.settings_menu_item),
@@ -34,8 +34,8 @@ fun SettingsScreen(
       isDarkTheme = uiState.appliedThemeOption == ThemeModeValue.DARK,
       modifier = Modifier.fillMaxSize(),
       onDarkThemeChange = { isDark ->
-        settingsViewModel.updateThemeOption(if (isDark) ThemeModeValue.DARK else ThemeModeValue.LIGHT)
-        settingsViewModel.saveThemeOption()
+        prefsViewModel.updateThemeOption(if (isDark) ThemeModeValue.DARK else ThemeModeValue.LIGHT)
+        prefsViewModel.saveThemeOption()
       },
       onNavigateToLanguageScreen = { navController.navigateTo(NavRoute.LanguageSetting) },
       onNavigateToPrivacyScreen = { navController.navigateTo(NavRoute.PrivacyPolicy) },